@@ -1,6 +1,6 @@
-const hexashades = require("../../dist/index.cjs.js");
+import { Colors } from '../../src/index'
 
-const ctx = new hexashades.Colors();
+const ctx = new Colors();
 
 describe("throws an error", () => {
   test("for invalid hex codes", () => {
@@ -43,24 +43,29 @@ describe("throws an error", () => {
   });
 
   test("for invalid percentage type", () => {
+    // @ts-expect-error Test for invalid percentage type
     expect(() => ctx.createColors("66", 'abc')).toThrow(
       "Invalid input. Wrong input types are given."
     );
+    // @ts-expect-error Test for invalid percentage type
     expect(() => ctx.createColors("66", true)).toThrow(
       "Invalid input. Wrong input types are given."
     );
   });
 
   test("for invalid hex code type", () => {
+    // @ts-expect-error Test for invalid hex code type: number
     expect(() => ctx.createColors(666, 10)).toThrow(
       "Invalid input. Wrong input types are given."
     );
+    // @ts-expect-error Test for invalid hex code type: Boolean
     expect(() => ctx.createColors(true, 10)).toThrow(
       "Invalid input. Wrong input types are given."
     );
   });
 
   test("for invalid prefix type", () => {
+    // @ts-expect-error Test for invalid prefix type
     expect(() => ctx.createColors('fff', 10, 'abc')).toThrow(
       "Invalid input. Wrong input types are given."
     );
